wip: before major rebuild
